+++ Dockerfile
@@ -1,8 +1,8 @@
 FROM golang:1.23-alpine AS builder
 COPY . .
-RUN CGO_ENABLED=0 GO111MODULE=on go build -a -installsuffix nocgo -o /thor ./cmd/thor
+RUN CGO_ENABLED=0 GO111MODULE=on go build -a -installsuffix nocgo -o /portpatrol ./cmd/portpatrol
 
 FROM scratch
-COPY --from=builder /thor /thor
-ENTRYPOINT ["/thor"]
+COPY --from=builder /portpatrol /portpatrol
+ENTRYPOINT ["/portpatrol"]
