+++ internal/checker/checker.go
