+++ internal/checker/checker_test.go
