+++ internal/checker/utils.go
