+++ internal/checker/utils_test.go
